script.gtv.helper/requirements.txt
requests>=2.22.0
xbmcgui
xbmcaddon
xbmcvfs